Add files via upload
In the past decade technology has come up a great way, Science Friction once is now a reality and available in every Household. Once we all must have wondered how our world is moving towards the immersive reality. We are continuously finding various ways and methods to interact with our surrounding using input devices. These input devices are used in giving commands for various operations to the PC, used for gaming, entrainment and many other such applications. Our gaming geeks must have enjoyed Nintendo, a gaming company that sold out the idea of 3D interactive method to interact with their consoles with the help of a handheld controller known as Wii Controller. It uses accelerometer to locate our gesture and sends the system. We have built a prototype which uses the accelerometer to read the gesture process it in a microcontroller and send it to the PC.
The required code is given above it is compiled with os.mbed compiler with USB, TSI and accelerometer being added to it. This code is then complied and uploaded to KL25Z.